update demo seeder

+++ database/seeders/tenant/demo/DemoSeeder.php
@@ -3,13 +3,15 @@
 namespace Database\Seeders\tenant\demo;
 
 use Carbon\Carbon;
+use App\Enums\PriorityLevel;
 use App\Models\LocationType;
-use App\Models\Tenants\Room;
 // use Illuminate\Database\Console\Seeds\WithoutModelEvents;
+use App\Models\Tenants\Room;
 use App\Models\Tenants\Site;
 use App\Models\Tenants\User;
 use App\Models\Tenants\Asset;
 use App\Models\Tenants\Floor;
+use App\Models\Tenants\Ticket;
 use App\Enums\NoticePeriodEnum;
 use App\Models\Tenants\Company;
 use App\Services\QRCodeService;
@@ -19,15 +21,20 @@ use App\Models\Tenants\Building;
 use App\Models\Tenants\Contract;
 use App\Models\Tenants\Provider;
 use App\Enums\ContractStatusEnum;
+use App\Enums\InterventionStatus;
 use Illuminate\Support\Facades\DB;
 use App\Enums\ContractDurationEnum;
 use App\Enums\MaintenanceFrequency;
 use App\Models\Central\CategoryType;
+use App\Models\Tenants\Intervention;
 use App\Models\Tenants\Maintainable;
+use App\Models\Tenants\MeterReading;
 use Illuminate\Support\Facades\Hash;
 use App\Enums\ContractRenewalTypesEnum;
-use App\Models\Tenants\MeterReading;
+use App\Models\Tenants\InterventionAction;
 use App\Models\Tenants\ScheduledNotification;
+use Database\Seeders\tenant\demo\ITDemoSeeder;
+use Database\Seeders\tenant\demo\HvacDemoSeeder;
 use App\Services\UserNotificationPreferenceService;
 use Database\Seeders\tenant\demo\VehicleDemoSeeder;
 use Database\Seeders\tenant\demo\ContractDemoSeeder;
@@ -40,6 +47,9 @@ class DemoSeeder extends Seeder
      */
     public function run(): void
     {
+        $tenant = tenancy()->tenant;
+        if ($tenant->id !== 'demo')
+            return;
 
         DB::statement('SET FOREIGN_KEY_CHECKS=0;');
         User::truncate();
@@ -47,20 +57,24 @@ class DemoSeeder extends Seeder
         DB::statement("TRUNCATE TABLE `provider_maintainable`");
         DB::statement("TRUNCATE TABLE `category_type_provider`");
         DB::statement("TRUNCATE TABLE `model_has_roles`");
+        DB::statement("TRUNCATE TABLE `contractables`");
         ScheduledNotification::truncate();
         Site::truncate();
         Building::truncate();
         Provider::truncate();
         Floor::truncate();
+        Ticket::truncate();
         MeterReading::truncate();
         Room::truncate();
         Asset::truncate();
         Contract::truncate();
+        Intervention::truncate();
+        InterventionAction::truncate();
 
         DB::statement('SET FOREIGN_KEY_CHECKS=1;');
 
 
-        $tenant = tenancy()->tenant;
+
 
         if (!User::where('email', 'super@sme-facility.com')->first()) {
 
@@ -104,7 +118,7 @@ class DemoSeeder extends Seeder
         $wallMaterials = CategoryType::where('category', 'wall_materials')->get();
         $floorMaterials = CategoryType::where('category', 'floor_materials')->get();
 
-        $site = Site::factory()->withMaintainableData([
+        Site::factory()->withMaintainableData([
             'name' => 'Site principal',
             'description' => 'Site de la démonstration',
 
@@ -114,14 +128,14 @@ class DemoSeeder extends Seeder
             'floor_material_id' => $floorMaterials[rand(0, count($floorMaterials) - 1)]->id,
             'wall_material_id' => $wallMaterials[rand(0, count($wallMaterials) - 1)]->id
         ]);
-        $building = Building::factory()->withMaintainableData(['name' => 'Bâtiment principal', 'description' => 'Bâtiment administratif'])->create();
+
+        Building::factory()->withMaintainableData(['name' => 'Bâtiment principal', 'description' => 'Bâtiment administratif'])->create();
 
         $floorGround = LocationType::where('level', 'floor')->where('slug', 'ground-floor')->first();
         $floorFloors = LocationType::where('level', 'floor')->where('slug', 'floors')->first();
         $floorGround = Floor::factory()->withMaintainableData([
             'name' => 'Rez-de-chaussée',
             'description' => 'Niveau 0',
-
         ])->create([
             'location_type_id' => $floorGround->id,
             'surface_floor' => 225.0,
@@ -129,6 +143,7 @@ class DemoSeeder extends Seeder
             'floor_material_id' => $floorMaterials[rand(0, count($floorMaterials) - 1)]->id,
             'wall_material_id' => $wallMaterials[rand(0, count($wallMaterials) - 1)]->id
         ]);
+
         $floorOne = Floor::factory()->withMaintainableData([
             'name' => 'Etage 1',
             'description' => 'Etage direction',
@@ -145,7 +160,7 @@ class DemoSeeder extends Seeder
         $roomTechnicalType = LocationType::where('level', 'room')->where('slug', 'technical-room')->first();
         $roomMeetingType = LocationType::where('level', 'room')->where('slug', 'meeting-room')->first();
 
-        $roomOfficeSales = Room::factory()->withMaintainableData(['name' => 'Bureau vente', 'description' => 'Bureau des commerciaux'])->create([
+        Room::factory()->withMaintainableData(['name' => 'Bureau vente', 'description' => 'Bureau des commerciaux'])->create([
             'level_id' => $floorGround->id,
             'location_type_id' => $roomOfficeType->id,
             'surface_floor' => 105.0,
@@ -155,7 +170,7 @@ class DemoSeeder extends Seeder
             'wall_material_id' => $wallMaterials[rand(0, count($wallMaterials) - 1)]->id
         ]);
 
-        $roomTechnical = Room::factory()->withMaintainableData([
+        Room::factory()->withMaintainableData([
             'name' => 'Local technique',
             'description' => 'Local chaufferie',
 
@@ -192,24 +207,6 @@ class DemoSeeder extends Seeder
             ]);
 
 
-        $provider = Provider::where('name', 'All Clean sa')->first();
-        $contract = Contract::factory()->create([
-            'provider_id' => $provider->id,
-            'name' => 'Nettoyage, entretien des locaux',
-            'type' => ContractTypesEnum::CLEANING->value,
-            'internal_reference' => 'CLEAN_2025-12',
-            'provider_reference' => fake()->randomLetter() . fake()->randomNumber(4, true),
-            'start_date' => Carbon::createFromDate(2025, 01, 12),
-            'contract_duration' => ContractDurationEnum::ONE_YEAR->value,
-            'end_date' => Carbon::createFromDate(2025, 01, 12)->addYear(),
-            'notice_period' => NoticePeriodEnum::ONE_MONTH->value,
-            'notice_date' => NoticePeriodEnum::ONE_MONTH->subFrom(Carbon::createFromDate(2025, 01, 12)->addYear()),
-            'renewal_type' => ContractRenewalTypesEnum::AUTOMATIC,
-            'status' => ContractStatusEnum::ACTIVE,
-            'notes' => fake()->text(50),
-        ]);
-
-        $roomOfficeDirector->contracts()->attach($contract);
 
         $roomMeeting = Room::factory()->withMaintainableData([
             'name' => 'Salle de réunion Einstein',
@@ -225,164 +222,7 @@ class DemoSeeder extends Seeder
             'height' => 2.50
         ]);
 
-        // Ordinateurs
-        $itGuy = User::factory()->withRole('Maintenance Manager')->create(['first_name' => 'Michael', 'last_name' => 'Durand', 'email' => 'michel.durand@sme-facility.com', 'job_position' => 'IT', 'phone_number' => '+3243764376', 'can_login' => true]);
-        app(UserNotificationPreferenceService::class)->createDefaultUserNotificationPreferences($itGuy);
 
-
-        $computerCategory = CategoryType::where('category', 'asset')->where('slug', 'asset-computer-hardware')->first();
-        $assetComputerSales = Asset::factory()
-            ->withMaintainableData(
-                [
-                    'name' => 'PC de Bureau HP commercial interne',
-                    'description' => 'PC de bureau pour les commerciaux internes',
-                    'purchase_date' => Carbon::now()->subYear(),
-                    'purchase_cost' => 899.99,
-                    'under_warranty' => true,
-                    'end_warranty_date' => Carbon::now()->addYear(2),
-                    'maintenance_manager_id' => $itGuy->id,
-                ]
-            )
-            ->forLocation($roomOfficeSales)
-            ->create([
-                'category_type_id' => $computerCategory->id,
-                'brand' => 'HP',
-                'model' => 'Pavilion H25B',
-                'serial_number' => 'X25-ABC-96',
-                "depreciable" => true,
-                "depreciation_start_date" => Carbon::now()->subYear(),
-                "depreciation_end_date" => Carbon::now()->addYear(3),
-                "depreciation_duration" =>  3,
-                "surface" => null
-            ]);
-
-        $assetComputerSales->refresh();
-        $assetComputerSales->maintainable->providers()->sync([Provider::where('name', 'Le comptoir de la ram')->first()->id]);
-
-        $assetComputerDirector = Asset::factory()
-            ->withMaintainableData(
-                [
-                    'name' => 'Laptop directeur ASUS',
-                    'description' => 'PC portable du directeur',
-                    'purchase_date' => Carbon::yesterday(),
-                    'purchase_cost' => 1299.99,
-                    'under_warranty' => true,
-                    'end_warranty_date' => Carbon::yesterday()->addYear(2),
-                    'maintenance_manager_id' => $itGuy->id,
-                ]
-            )
-            ->forLocation($roomOfficeDirector)
-            ->create([
-                'category_type_id' => $computerCategory->id,
-                'brand' => 'Asus',
-                'model' => 'Vivobook E25F',
-                'serial_number' => 'AZ5-CD-257BC',
-                "depreciable" => true,
-                "depreciation_start_date" => Carbon::yesterday(),
-                "depreciation_end_date" => Carbon::yesterday()->addYear(2),
-                "depreciation_duration" =>  2,
-                "surface" => null
-            ]);
-
-        $assetComputerDirector->refresh();
-        $assetComputerDirector->maintainable->providers()->sync([Provider::where('name', 'Le comptoir de la ram')->first()->id]);
-
-
-        $assetSoftDeleted = Asset::factory()
-            ->withMaintainableData(
-                [
-                    'name' => 'PC de Bureau HP de la secrétaire',
-                    'description' => 'PC de bureau à l\'accueil',
-                    'purchase_date' => Carbon::yesterday()->subYears(4),
-                    'purchase_cost' => 599.99,
-                    'under_warranty' => false,
-                    'end_warranty_date' => Carbon::yesterday()->subYears(2),
-                    'maintenance_manager_id' => $itGuy->id,
-                ]
-            )
-            ->forLocation($roomOfficeSales)
-            ->create([
-                'category_type_id' => $computerCategory->id,
-                'brand' => 'HP',
-                'model' => 'Pavilion H25B',
-                'serial_number' => 'X25-ABC-96',
-                "depreciable" => false,
-                "depreciation_start_date" => null,
-                "depreciation_end_date" => null,
-                "depreciation_duration" =>  null,
-                "surface" => null,
-                "deleted_at" => Carbon::yesterday()->subYears(2),
-            ]);
-
-        $assetSoftDeleted->refresh();
-        $assetSoftDeleted->maintainable->providers()->sync([Provider::where('name', 'Le comptoir de la ram')->first()->id]);
-
-
-        $provider = Provider::where('name', 'Le comptoir de la ram')->first();
-        $contract = Contract::factory()->create([
-            'provider_id' => $provider->id,
-            'name' => 'Contrat de maintenance IT',
-            'type' => ContractTypesEnum::ONDEMAND->value,
-            'internal_reference' => 'PC Repair',
-            'provider_reference' => fake()->randomLetter() . fake()->randomNumber(4, true),
-            'start_date' => Carbon::yesterday()->subYears(2),
-            'contract_duration' => ContractDurationEnum::SIX_MONTHS->value,
-            'end_date' => ContractDurationEnum::SIX_MONTHS->addTo(Carbon::yesterday()->subYears(2)),
-            'notice_period' => null,
-            'notice_date' => null,
-            'renewal_type' => ContractRenewalTypesEnum::AUTOMATIC,
-            'status' => ContractStatusEnum::CANCELLED,
-            'notes' => fake()->text(50),
-        ]);
-
-        $assetComputerSales->contracts()->attach($contract);
-        $assetComputerDirector->contracts()->attach($contract);
-
-        // HVAC
-        $hvacCategory = CategoryType::where('category', 'asset')->where('slug', 'asset-hvac')->first();
-        $assetHvac = Asset::factory()
-            ->withMaintainableData(
-                [
-                    'name' => 'Chaudière gaz',
-                    'description' => 'Chaudière gaz Frisquet',
-                    'need_maintenance' => true,
-                    'maintenance_frequency' => MaintenanceFrequency::ANNUAL->value,
-                    'last_maintenance_date' => Carbon::now()->subYear(),
-                    'next_maintenance_date' => Carbon::now()->tomorrow(),
-                ]
-            )
-            ->forLocation($roomTechnical)
-            ->create([
-                'category_type_id' => $hvacCategory->id,
-                'brand' => 'Frisquet',
-                'model' => 'HYDROCONFORT',
-                'serial_number' => '15869AD44PLD',
-                'surface' => null,
-                'has_meter_readings' => false,
-                "depreciable" => false,
-            ]);
-
-        $assetHvac->refresh();
-        $hvacProvider = Provider::where('name', 'Le comptoir du froid')->first();
-        $assetHvac->maintainable->providers()->sync([$hvacProvider->id]);
-
-        $contract = Contract::factory()->create([
-            'provider_id' => $hvacProvider->id,
-            'name' => 'Contrat de maintenance HVAC',
-            'type' => ContractTypesEnum::MAINTENANCE->value,
-            'internal_reference' => 'HVAC_MAINTENANCE',
-            'provider_reference' => fake()->randomLetter() . fake()->randomNumber(4, true),
-            'start_date' => Carbon::now()->subYear(),
-            'contract_duration' => ContractDurationEnum::ONE_YEAR->value,
-            'end_date' => ContractDurationEnum::ONE_YEAR->addTo(Carbon::now()->subYear()),
-            'notice_period' => NoticePeriodEnum::FOURTEEN_DAYS->value,
-            'notice_date' => NoticePeriodEnum::FOURTEEN_DAYS->subFrom(ContractDurationEnum::ONE_YEAR->addTo(Carbon::now()->subYear())),
-            'renewal_type' => ContractRenewalTypesEnum::AUTOMATIC,
-            'status' => ContractStatusEnum::ACTIVE,
-            'notes' => fake()->text(50),
-        ]);
-
-        $assetHvac->contracts()->attach($contract);
 
         // Luminaire salle de réunion
         $lightingCategory = CategoryType::where('category', 'asset')->where('slug', 'asset-lighting')->first();
@@ -400,33 +240,57 @@ class DemoSeeder extends Seeder
                 'model' => 'Maxi-Lamp',
             ]);
 
-        app(QRCodeService::class)->createAndAttachQR($assetComputerSales);
-        app(QRCodeService::class)->createAndAttachQR($assetComputerDirector);
 
-        app(QRCodeService::class)->createAndAttachQR($assetHvac);
         app(QRCodeService::class)->createAndAttachQR($assetLighting);
 
-        $lightingCategory = CategoryType::where('category', 'asset')->where('slug', 'asset-lighting')->first();
+        $ticket = Ticket::factory()->forLocation($assetLighting)->create(['description' => 'Ampoule ne fonctionne plus', 'reported_by' => $admin->id, 'created_at' => Carbon::yesterday()]);
 
-        $assetLighting = Asset::factory()
+        $interventionRepair = CategoryType::where('category', 'intervention')->where('slug', 'intervention-repair')->first();
+        $intervention = Intervention::factory()->forTicket($ticket)->create([
+            'description' => '2 ampoules à remplacer',
+            'intervention_type_id' => $interventionRepair->id,
+            'priority' => PriorityLevel::URGENT->value,
+            'status' => InterventionStatus::COMPLETED->value,
+            'planned_at' => Carbon::now(),
+            'repair_delay' => null,
+        ]);
+
+        $actionType = CategoryType::where('category', 'action')->where('slug', 'action-repair')->first();
+        InterventionAction::factory()->forIntervention($intervention)->create([
+            'action_type_id' => $actionType->id,
+            'description' => 'Ampoules changées',
+            'intervention_date' => Carbon::now(),
+            'started_at' => '09:30',
+            'finished_at' => '09:40',
+            'intervention_costs' => 0.0,
+            'creator_email' => fake()->safeEmail()
+        ]);
+
+        $assetLightingHalo = Asset::factory()
             ->withMaintainableData(
                 [
                     'name' => 'Lampe plafond',
-                    'description' => 'Lampe plafond - Socket GU10',
+                    'description' => 'Lampe plafond - Halogène',
                 ]
             )
-            ->forLocation($roomMeeting)
+            ->forLocation($floorGround)
             ->create([
                 'category_type_id' => $lightingCategory->id,
                 'brand' => 'Illudesign',
-                'model' => 'Maxi-Lamp',
+                'model' => 'Lamp-Halo',
             ]);
 
-        $assetLighting->refresh();
-        $assetLighting->maintainable->providers()->sync([Provider::where('name', 'Jacky Den SPRL')->first()->id]);
+        $assetLightingHalo->refresh();
+        $assetLightingHalo->maintainable->providers()->sync([Provider::where('name', 'Jacky Den SPRL')->first()->id]);
+
+        $ticket = Ticket::factory()->forLocation($assetLightingHalo)->create(['description' => 'Ampoule clignote', 'reported_by' => $admin->id, 'created_at' => Carbon::yesterday()]);
+
+        app(QRCodeService::class)->createAndAttachQR($assetLightingHalo);
 
         $this->call([
             VehicleDemoSeeder::class,
+            ITDemoSeeder::class,
+            HvacDemoSeeder::class,
             ContractDemoSeeder::class
         ]);
     }
